% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_figures.R
\name{freq_ontologies}
\alias{freq_ontologies}
\title{freq_ontologies()}
\usage{
freq_ontologies(X, type = c("composition", "class", "both"))
}
\arguments{
\item{X}{can be either a character vector containing Nglycan names OR a list of multiple character vectors, the name of the list components will be use to generate the legend}

\item{type}{can be either <composition>,<class>,or <both> to indicate the type of ontology terms to be considered in the figure}
}
\value{
a barplot showing the ontology of the selected
}
\description{
This function will read one or more Nglycan_lists and generate frequency plots for the N-Glycan components
}
\author{
Geremy Clair
}
