% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base_functions.R
\docType{data}
\name{Nglycan_db}
\alias{Nglycan_db}
\title{Nglycan_db}
\format{
A character
}
\usage{
Nglycan_db
}
\description{
This character vector contains the list of all the Nglycans currently present in Metaspace (as of 07/12/24).
}
\keyword{datasets}
