% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base_functions.R
\name{Nglycan_KS}
\alias{Nglycan_KS}
\title{Nglycan_KS()}
\usage{
Nglycan_KS(rankingTable, order = "ascending")
}
\arguments{
\item{rankingTable}{should be a two column data.frame containing the Nglycan names as first column and the ranking values as second column}

\item{order}{should either be "ascending" or "descending" to indicate the order of the ranking values to use.}
}
\value{
a data.frame containing enrichment results from the GO retrieve from UniProt.
}
\description{
Allows to perform Nglycan Ontology enrichment using a Kolmogorov-Smirnov test, in this case, the "rank" of the Nglycan will be used to perform the enrichment the rank can be directly provided or derived from a list of pvalues or fold changes.
}
\author{
Geremy Clair
}
