% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base_functions.R
\name{NGlycan_miner}
\alias{NGlycan_miner}
\title{NGlycan_miner()}
\usage{
NGlycan_miner(X)
}
\arguments{
\item{X}{a character vector with each element being the name of a given N-glycan.}
}
\value{
This function returns a table providing the appartenance of N glycan to different glycan ontology terms.
}
\description{
This function will read a list of N_glycan names located in a character vector (each glycan is one element of the char_vector) the algorithm contained in this function will parse the information contained in the glycan names to define the appartenance of the different glycan to glycan ontology terms.
}
\author{
Dusan Velickovic, Geremy Clair
}
