% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_base_functions.R
\name{NGlycan_ontologies}
\alias{NGlycan_ontologies}
\title{NGlycan_ontologies()}
\usage{
NGlycan_ontologies(X)
}
\arguments{
\item{X}{a character vector with each element being the name of a given N-glycan.}
}
\value{
This function returns a list of ontology terms associated with the different NGlycans in the list and provide their association with the N-Glycans
}
\description{
This function will generate ntologies from a list of N-Glycan Names. To get the results in a tabular format instead of a list of ontology terms, please use the function NGlycan_miner
}
\author{
Dusan Velickovic, Geremy Clair
}
